Take exceptions into separate files
